+++ db/default_song_library.sql
@@ -1,3 +1,91 @@
-INSERT INTO bot_default_song_library (id, artist, song_title, album, length_in_seconds) VALUES
-(1, "test1", "test1", "test1", 5),
-(2, "test2", "test2", "test2", 5);
+INSERT INTO bot_default_song_library (artist, song_title, album, length_in_seconds) VALUES
+/* porter robinson */
+-- pre-spitfire
+("Porter Robinson", "Say My Name", "Say My Name", 392),
+-- spitfire
+("Porter Robinson", "Spitfire", "Spitfire", 405),
+("Porter Robinson", "Unison", "Spitfire", 350),
+("Porter Robinson", "100% in the Bitch", "Spitfire", 251),
+("Porter Robinson", "Vandalism (ft. Amba Shepherd)", "Spitfire", 440),
+("Porter Robinson", "The State", "Spitfire", 357),
+("Porter Robinson", "The Seconds (ft. Jano)", "Spitfire", 343),
+-- language + easy
+("Porter Robinson", "Language", "Language", 368),
+("Mat Zo & Porter Robinson", "Easy", "Easy", 396),
+-- worlds
+("Porter Robinson", "Divinity (ft. Amy Millan)", "Worlds", 368),
+("Porter Robinson", "Sad Machine", "Worlds", 350),
+("Porter Robinson", "Years of War (ft. Breanne Düren and Sean Caskey)", "Worlds", 236),
+("Porter Robinson", "Flicker", "Worlds", 279),
+("Porter Robinson", "Fresh Static Snow", "Worlds", 358),
+("Porter Robinson", "Polygon Dust (ft. Lemaitre)", "Worlds", 209),
+("Porter Robinson", "Hear The Bells (ft. Imaginary Cities)", "Worlds", 286),
+("Porter Robinson", "Natural Light", "Worlds", 178),
+("Porter Robinson", "Lionhearted (ft. Urban Cone)", "Worlds", 264),
+("Porter Robinson", "Sea of Voices", "Worlds", 299),
+("Porter Robinson", "Fellow Feeling", "Worlds", 349),
+("Porter Robinson", "Goodbye to a World", "Worlds", 328),
+-- shepherdess
+("Porter Robinson", "Shepherdess", "Lionhearted Vinyl", 436),
+-- virtual self
+("VIRTUAL SELF", "Particle Arts", "VIRTUAL SELF", 234),
+("VIRTUAL SELF", "Ghost Voices", "VIRTUAL SELF", 264),
+("VIRTUAL SELF", "a.i.ngel (Become God)", "VIRTUAL SELF", 299),
+("VIRTUAL SELF", "Key", "VIRTUAL SELF", 328),
+("VIRTUAL SELF", "EON BREAK", "VIRTUAL SELF", 220),
+("VIRTUAL SELF", "ANGEL VOICES", "ANGEL VOICES", 392),
+-- nurture
+("Porter Robinson", "Lifelike", "Nurture", 95),
+("Porter Robinson", "Look At The Sky", "Nurture", 310),
+("Porter Robinson", "Get Your Wish", "Nurture", 218),
+("Porter Robinson", "Wind Tempos", "Nurture", 364),
+("Porter Robinson", "Musician", "Nurture", 238),
+("Porter Robinson", "do-re-mi-fa-so-la-ti-do", "Nurture", 215),
+("Porter Robinson", "Mother", "Nurture", 216),
+("Porter Robinson", "dullscythe", "Nurture", 240),
+("Porter Robinson", "Sweet Time", "Nurture", 252),
+("Porter Robinson", "Mirror", "Nurture", 307),
+("Porter Robinson", "Something Comforting", "Nurture", 282),
+("Porter Robinson", "Blossom", "Nurture", 226),
+("Porter Robinson & Totally Enormous Extinct Dinosaurs", "Unfold", "Nurture", 286),
+("Porter Robinson", "Trying to Feel Alive", "Nurture", 284),
+("Porter Robinson & Wednesday Campanella", "fullmoon lullaby", "Nurture", 243),
+-- post-nurture
+("Porter Robinson & League of Legends", "Everything Goes On", "Everything Goes On", 202),
+/* porter robinson & madeon */
+-- shelter
+("Porter Robinson & Madeon", "Shelter", "Shelter", 219),
+/* madeon */
+-- adventure (deluxe)
+("Madeon", "Isometric (Intro)", "Adventure (Deluxe)", 80),
+("Madeon", "You're On (ft. Kyan)", "Adventure (Deluxe)", 192),
+("Madeon", "OK", "Adventure (Deluxe)", 181),
+("Madeon", "La Lune (ft. Dan Smith)", "Adventure (Deluxe)", 219),
+("Madeon", "Pay No Mind (ft. Passion Pit)", "Adventure (Deluxe)", 249),
+("Madeon", "Beings", "Adventure (Deluxe)", 215),
+("Madeon", "Imperium", "Adventure (Deluxe)", 218),
+("Madeon", "Zephyr", "Adventure (Deluxe)", 220),
+("Madeon", "Nonsense (ft. Mark Foster)", "Adventure (Deluxe)", 225),
+("Madeon", "Innocence (ft. Aquilo)", "Adventure (Deluxe)", 224),
+("Madeon", "Pixel Empire", "Adventure (Deluxe)", 244),
+("Madeon", "Home", "Adventure (Deluxe)", 225),
+("Madeon", "Icarus", "Adventure (Deluxe)", 214),
+("Madeon", "Finale (ft. Nicholas Petricca)", "Adventure (Deluxe)", 204),
+("Madeon", "The City", "Adventure (Deluxe)", 234),
+("Madeon", "Cut the Kid", "Adventure (Deluxe)", 197),
+("Madeon", "Technicolor", "Adventure (Deluxe)", 385),
+("Madeon", "Only Way Out (ft. Vancouver Sleep Clinic)", "Adventure (Deluxe)", 226),
+-- good faith era
+("Madeon", "Dream Dream Dream", "Good Faith", 234),
+("Madeon", "All My Friends", "Good Faith", 204),
+("Madeon", "Be Fine", "Good Faith", 208),
+("Madeon", "Nirvana", "Good Faith", 152),
+("Madeon", "Mania", "Good Faith", 152),
+("Madeon", "Miracle", "Good Faith", 250),
+("Madeon", "No Fear No More", "Good Faith", 195),
+("Madeon", "Hold Me Just Because", "Good Faith", 186),
+("Madeon", "Heavy With Hoping", "Good Faith", 241),
+("Madeon", "Borealis", "Good Faith", 285),
+("Madeon", "The Prince", "The Prince", 222),
+("Madeon", "Love You Back", "Love You Back", 214),
+("Madeon", "Gonna Be Good", "Gonna Be Good (Length to fix at release)", 295);